requirements.txt
gql
python-dotenv